Narrowing conversion fixes

+++ src/gnb/gtp/utils.cpp
@@ -78,16 +78,16 @@ TokenBucket::TokenBucket(long byteCapacity) : byteCapacity(byteCapacity)
     }
 }
 
-bool TokenBucket::tryConsume(int64_t numberTokens)
+bool TokenBucket::tryConsume(uint64_t numberTokens)
 {
     if (byteCapacity > 0)
     {
         refill();
-        if (availableTokens < numberTokens)
+        if (availableTokens < static_cast<double>(numberTokens))
             return false;
         else
         {
-            availableTokens -= numberTokens;
+            availableTokens -= static_cast<double>(numberTokens);
             return true;
         }
     }
@@ -95,7 +95,7 @@ bool TokenBucket::tryConsume(int64_t numberTokens)
         return true;
 }
 
-void TokenBucket::updateCapacity(int64_t newByteCapacity)
+void TokenBucket::updateCapacity(uint64_t newByteCapacity)
 {
     byteCapacity = newByteCapacity;
     if (newByteCapacity > 0)
@@ -108,13 +108,13 @@ void TokenBucket::refill()
     if (currentTimeMillis > lastRefillTimestamp)
     {
         int64_t millisSinceLastRefill = currentTimeMillis - lastRefillTimestamp;
-        double refill = millisSinceLastRefill * refillTokensPerOneMillis;
+        double refill = static_cast<double>(millisSinceLastRefill) * refillTokensPerOneMillis;
         availableTokens = std::min(static_cast<double>(byteCapacity), availableTokens + refill);
         lastRefillTimestamp = currentTimeMillis;
     }
 }
 
-bool RateLimiter::allowDownlinkPacket(uint64_t pduSession, int64_t packetSize)
+bool RateLimiter::allowDownlinkPacket(uint64_t pduSession, uint64_t packetSize)
 {
     int ueId = GetUeId(pduSession);
 
@@ -135,7 +135,7 @@ bool RateLimiter::allowDownlinkPacket(uint64_t pduSession, int64_t packetSize)
     return true;
 }
 
-bool RateLimiter::allowUplinkPacket(uint64_t pduSession, int64_t packetSize)
+bool RateLimiter::allowUplinkPacket(uint64_t pduSession, uint64_t packetSize)
 {
     int ueId = GetUeId(pduSession);
 
@@ -156,7 +156,7 @@ bool RateLimiter::allowUplinkPacket(uint64_t pduSession, int64_t packetSize)
     return true;
 }
 
-void RateLimiter::updateUeUplinkLimit(int ueId, int64_t limit)
+void RateLimiter::updateUeUplinkLimit(int ueId, uint64_t limit)
 {
     if (limit <= 0)
     {
@@ -175,7 +175,7 @@ void RateLimiter::updateUeUplinkLimit(int ueId, int64_t limit)
     }
 }
 
-void RateLimiter::updateUeDownlinkLimit(int ueId, int64_t limit)
+void RateLimiter::updateUeDownlinkLimit(int ueId, uint64_t limit)
 {
     if (limit <= 0)
     {
@@ -194,7 +194,7 @@ void RateLimiter::updateUeDownlinkLimit(int ueId, int64_t limit)
     }
 }
 
-void RateLimiter::updateSessionUplinkLimit(uint64_t pduSession, int64_t limit)
+void RateLimiter::updateSessionUplinkLimit(uint64_t pduSession, uint64_t limit)
 {
     if (limit <= 0)
     {
@@ -213,7 +213,7 @@ void RateLimiter::updateSessionUplinkLimit(uint64_t pduSession, int64_t limit)
     }
 }
 
-void RateLimiter::updateSessionDownlinkLimit(uint64_t pduSession, int64_t limit)
+void RateLimiter::updateSessionDownlinkLimit(uint64_t pduSession, uint64_t limit)
 {
     if (limit <= 0)
     {
